requirements.txt
aiogram~=2.21
pytube==12.1.0
Pillow==9.2.0